Support parallel test execution.
Executing test can take some time. In order to improve the performance
the parallelis was introduced in the application. It is very
important that the parallel execution doesn't affect the measured
times. In order to achieve this goal only CPU - 1 is used and the
left one should be enough to support the operating system.